renamed project using proper java/mvn naming conventions
