+++ src/test/groovy/com/samplelogin/geb/spock/SampleLoginPageSpec.groovy
@@ -1,8 +1,7 @@
-package com.gebspockproject.template
+package com.samplelogin.geb.spock
 
-
-import com.gebspockproject.template.pages.SampleLoginPage
-import com.gebspockproject.template.targetpages.TargetPageType
+import com.samplelogin.geb.spock.targetpages.TargetPageType
+import com.samplelogin.geb.spock.pages.SampleLoginPage
 import spock.lang.Unroll
