+++ src/test/groovy/com/samplelogin/geb/spock/annotations/OnDemand.groovy
@@ -1,4 +1,4 @@
-package com.gebspockproject.template.annotations
+package com.samplelogin.geb.spock.annotations
 
 import java.lang.annotation.ElementType
 import java.lang.annotation.Retention
@@ -6,14 +6,13 @@ import java.lang.annotation.RetentionPolicy
 import java.lang.annotation.Target
 
 
-
 /*
-  The EndToEnd annotation (tag) indicates that the annotated Spock Test
-  performs an EndToEnd Test.
+  The OnDemand annotation (tag) indicates that the annotated Spock Test
+  should only execute when explicitly specified
  */
 
 
 @Target([ElementType.TYPE, ElementType.METHOD])
 @Retention(RetentionPolicy.RUNTIME)
-public @interface  EndToEnd {
+public @interface  OnDemand {
 }
