+++ src/test/groovy/com/samplelogin/geb/spock/pages/SampleSecureAreaPage.groovy
@@ -1,4 +1,4 @@
-package com.gebspockproject.template.pages
+package com.samplelogin.geb.spock.pages
 
 import geb.Page
